requirements.txt
# Core
langchain>=0.3
langgraph>=0.2
langsmith>=0.1

# API
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.0
pydantic-settings>=2.0

# Queue
aio-pika>=9.0

# Database
sqlalchemy>=2.0
alembic>=1.13
asyncpg>=0.29
psycopg2-binary>=2.9

# WhatsApp
httpx>=0.27

# Media Processing
openai-whisper>=20231117
Pillow>=10.0
python-docx>=1.1
PyPDF2>=3.0
openpyxl>=3.1
python-magic>=0.4

# LLM
langchain-openai>=0.2

# Jobs
apscheduler>=3.10

# Knowledge Base / RAG
chromadb>=0.5
langchain-chroma>=0.1
langchain-text-splitters>=0.2

# Utils
python-dotenv>=1.0

# Auth
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.9
email-validator>=2.0
